Added more in() contracts and found a problem with range interface
The copy algorithm discovered a problem, when source and target range are
of the LinkedList class. The node is put() to the target before it is
removed from the source. This would mess up the linkage pointers.

So, direct copy isn't possible. But maybe some kind of adapter might do
the job here, which just takes one node out of the source and passed it
later.